Add err_chk and restore support for etcd_backup.

+++ etcd_backup.sh
@@ -23,13 +23,30 @@ if [ "${etcd_servers%%https:*}" == "X" ];then
         echo "Error: got etcd endpoints with tls, but failed to find out the tls cert files."
         exit 0
     fi
-    /etcdctl --endpoints "${etcd_servers#X}" \
-        --cert=${etcd_certfile#X} \
-        --key=${etcd_keyfile#X} \
-        --cacert=${etcd_cafile#X} \
-        snapshot save /var/lib/etcd/etcd_v3_backup_`date +%Y%m%d-%H%M%S`
+
+    cmd="/etcdctl --endpoints ${etcd_servers#X} --cert=${etcd_certfile#X} --key=${etcd_keyfile#X} --cacert=${etcd_cafile#X}"
 else
-    /etcdctl --endpoints "${etcd_servers#X}" snapshot save /var/lib/etcd/etcd_v3_backup_`date +%Y%m%d-%H%M%S`
+    cmd="/etcdctl --endpoints ${etcd_servers#X}"
 fi
-cd /var/lib/etcd/
-ls -t etcd_v3_backup_* |sed '1,7d'|xargs -n1 rm -rf
+
+polit_v=`date +%s`
+${cmd} put "/fnt/polit" "${polit_v}"
+tvar=X`ETCDCTL_API=3 /etcdctl --endpoints="http://127.0.0.1:2379" get "/fnt/polit"|awk 'NR==2{print $1}'`
+if [ ! -d /var/lib/k8s_etcd_backup ];then
+    mkdir -p /var/lib/k8s_etcd_backup
+fi
+cd /var/lib/k8s_etcd_backup/
+if [ "${tvar#X}" == "${polit_v}" ];then
+    ${cmd} snapshot save /var/lib/k8s_etcd_backup/etcd_v3_backup_`date +%Y%m%d-%H%M%S`
+    ls -t etcd_v3_backup_* |sed '1,7d'|xargs -n1 rm -rf
+else
+    echo "Warning: detected error when tried to put&get etcd data."
+    latest_backup=X`ls etcd_v3_backup_* -td 2>/dev/null|awk 'NR==1{print $1}'`
+    if [ "${latest_backup}" == "X" ];then
+        echo "Error: there is no valid backup."
+        exit 0
+    fi
+    mv /var/lib/etcd /var/lib/broken_etcd_`date +%Y%m%d-%H%M%S`
+    echo "Restoring data from latest backup ${latest_backup#X}..."
+    ${cmd} snapshot restore /var/lib/k8s_etcd_backup/${latest_backup#X} --data-dir=/var/lib/etcd
+fi
